# docs/openapi-auto.yml
components:
  schemas:
...
    AnotacionCreate:
      additionalProperties: false
      properties:
        alumno_id:
          description: ID del alumno
          example: 1
          type: integer
        curso_id:
          description: ID del curso
          example: 1
          type: integer
        curso_profesor_id:
          description: ID de la relación Curso_Profesores
          example: 1
          type: integer
        inscripcion_id:
          description: ID de la inscripción del alumno
          example: 1
          type: integer
        sesion_id:
          description: ID de la sesión
          example: 1
          type: integer
        texto:
          description: Texto descriptivo de la anotación
          example: Excelente participación en la discusión sobre gramática inglesa.
            Demuestra comprensión avanzada.
          nullable: true
          type: string
        tipo_anotacion:
          description: 'Tipo de anotación permitido: Evaluacion, Comportamiento, Observacion,
            Otros (NO se permite Ausencia - usar pasar-lista)'
          enum:
          - Evaluacion
          - Comportamiento
          - Observacion
          - Otros
          example: Evaluacion
          type: string
      required:
      - alumno_id
      - curso_id
      - curso_profesor_id
      - inscripcion_id
      - sesion_id
      - tipo_anotacion
      type: object
    AnotacionUpdate:
      additionalProperties: false
      properties:
        alumno_id:
          nullable: true
          type: integer
          writeOnly: true
        curso_id:
          nullable: true
          type: integer
          writeOnly: true
        curso_profesor_id:
          nullable: true
          type: integer
          writeOnly: true
        inscripcion_id:
          nullable: true
          type: integer
          writeOnly: true
        motivo_baja:
          description: Motivo de baja (para dar de baja la anotación)
          example: Anotación errónea
          nullable: true
          type: string
        sesion_id:
          nullable: true
          type: integer
          writeOnly: true
        texto:
          description: Nuevo texto de la anotación
          example: Comportamiento disruptivo en clase
          nullable: true
          type: string
        timestamp_alta:
          format: date-time
          nullable: true
          type: string
          writeOnly: true
        timestamp_baja:
          format: date-time
          nullable: true
          type: string
          writeOnly: true
        tipo_anotacion:
          description: Nuevo tipo de anotación
          enum:
          - Ausencia
          - Evaluacion
          - Comportamiento
          - Observacion
          - Otros
          example: Comportamiento
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  description: Spec generated from code (partial)
  title: API Workers Profesores
  version: 1.0.0
openapi: 3.0.2
paths:
...
  /anotaciones/:
    get:
      description: Auto-generated operation for anotaciones.listar_anotaciones
      operationId: anotaciones.listar_anotaciones
      parameters:
      - in: query
        name: sesion_id
        required: false
        schema:
          type: string
      - in: query
        name: alumno_id
        required: false
        schema:
          type: string
      - in: query
        name: curso_id
        required: false
        schema:
          type: string
      - in: query
        name: tipo_anotacion
        required: false
        schema:
          type: string
      - in: query
        name: excluir_tipo
        required: false
        schema:
          type: string
      - in: query
        name: fecha_desde
        required: false
        schema:
          type: string
      - in: query
        name: fecha_hasta
        required: false
        schema:
          type: string
      - in: query
        name: activas
        required: false
        schema:
          type: string
      - in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: string
      - in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: string
      - in: query
        name: with_total
        required: false
        schema:
          type: string
      - in: query
        name: order_by
        required: false
        schema:
          type: string
      - in: query
        name: order_dir
        required: false
        schema:
          type: string
      - in: query
        name: expand
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - bearerAuth: []
      summary: GET /anotaciones/
      tags:
      - Anotaciones
    post:
      description: Auto-generated operation for anotaciones.crear_anotacion
      operationId: anotaciones.crear_anotacion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnotacionCreate'
        required: true
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - bearerAuth: []
      summary: POST /anotaciones/
      tags:
      - Anotaciones
  /anotaciones/{anotacion_id}:
    delete:
      description: Auto-generated operation for anotaciones.eliminar_anotacion
      operationId: anotaciones.eliminar_anotacion
      parameters:
      - in: path
        name: anotacion_id
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - bearerAuth: []
      summary: DELETE /anotaciones/{anotacion_id}
      tags:
      - Anotaciones
    get:
      description: Auto-generated operation for anotaciones.obtener_anotacion
      operationId: anotaciones.obtener_anotacion
      parameters:
      - in: path
        name: anotacion_id
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - bearerAuth: []
      summary: GET /anotaciones/{anotacion_id}
      tags:
      - Anotaciones
    patch:
      description: Auto-generated operation for anotaciones.actualizar_anotacion
      operationId: anotaciones.actualizar_anotacion
      parameters:
      - in: path
        name: anotacion_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnotacionUpdate'
        required: true
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - bearerAuth: []
      summary: PATCH /anotaciones/{anotacion_id}
      tags:
      - Anotaciones
    put:
      description: Auto-generated operation for anotaciones.actualizar_anotacion
      operationId: anotaciones.actualizar_anotacion
      parameters:
      - in: path
        name: anotacion_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnotacionUpdate'
        required: true
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - bearerAuth: []
      summary: PUT /anotaciones/{anotacion_id}
      tags:
      - Anotaciones
...